+++ packages/dartz_plus/lib/src/annotations.dart
@@ -15,7 +15,8 @@ class Mapper {
   ///
   /// [target] is the Type of the class you want to map TO.
   /// [reverse] if true, generates a reverse mapping method (Target -> Source).
-  const Mapper(this.target, {this.reverse = true});
+  /// [constructor] the name of the target constructor to use.
+  const Mapper(this.target, {this.reverse = true, this.constructor});
 
   /// The target type to map to.
   ///
@@ -24,4 +25,27 @@ class Mapper {
 
   /// Whether to generate the reverse mapping (Target -> Source).
   final bool reverse;
+
+  /// The name of the target class constructor to use.
+  /// If null, the default (unnamed) constructor is used.
+  final String? constructor;
+}
+
+///
+/// Annotation to map a source field to a target field with a different name.
+///
+class MapTo {
+  /// Creates a MapTo annotation.
+  const MapTo(this.name);
+
+  /// The name of the target field this field maps to.
+  final String name;
+}
+
+///
+/// Annotation to ignore a field during mapping.
+///
+class IgnoreMap {
+  /// Creates an IgnoreMap annotation.
+  const IgnoreMap();
 }